pyproject.toml
[project]
name = "agentic-connector-builder-webapp"
version = "0.1.0"
description = "A full-stack agentic connector builder webapp with YAML editing capabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "reflex>=0.8.17",
    "reflex-monaco",
    "pydantic-ai>=1.9.1",
    "pydantic-ai-slim[duckduckgo]",
    "python-dotenv>=1.2.1",
    "dill>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "playwright>=1.40.0",
]

[project.scripts]
agentic-connector-builder-webapp = "agentic_connector_builder_webapp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentic_connector_builder_webapp"]

[dependency-groups]
dev = [
    "playwright>=1.54.0",
    "poethepoet>=0.37.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.9",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
